change: rename files in Capacity/Components/FolderBar
