+++ src/pages/Capacity/Capacity.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import Carousel from './Components/Carousel/CustomCarousel';
-import PlaylistRecomendada from './Components/PlaylistBar/PlaylistBarInitial'
-import PlaylistSecundary from './Components/PlaylistBar/PlaylistBarSecundary'
-import PlaylistTertiary from './Components/PlaylistBar/PlaylistBarTertiary'
-import PlaylistBarOther from './Components/PlaylistBar/PlaylistBarOther'
+import PlaylistRecomendada from './Components/FolderBar/FolderBarInitial'
+import PlaylistSecundary from './Components/FolderBar/FolderBarSecundary'
+import PlaylistTertiary from './Components/FolderBar/FolderBarTertiary'
+import PlaylistBarOther from './Components/FolderBar/FolderBarOther'
 import AllCategories from './Components/ButtonCategory/ButtonCategory'
 import AnimationHome from './Components/Animation/VideoHomeCapacity'
 import NavbarCategory from './Components/NavbarSide/NavbarCategory'
-import AllFolders from './Components/PlaylistBar/AllPlaylists';
+import AllFolders from './Components/FolderBar/AllFolders';
 import Footer from '../../components/Footer/Footer';
 
 import './Capacity.css'
